+++ src/lib/nodes/Gateway.spec.ts
@@ -4,7 +4,7 @@ import { reSerializeCertificate } from '../_test_utils';
 import { generateRSAKeyPair, getPrivateAddressFromIdentityKey } from '../crypto_wrappers/keys';
 import Certificate from '../crypto_wrappers/x509/Certificate';
 import { MockKeyStoreSet } from '../keyStores/testMocks';
-import { issueGatewayCertificate } from '../pki';
+import { issueGatewayCertificate } from '../pki/issuance';
 import { Gateway } from './Gateway';
 import { StubVerifier } from './signatures/_test_utils';
