+++ src/lib/nodes/channels/Channel.spec.ts
@@ -9,7 +9,7 @@ import {
 } from '../../crypto_wrappers/keys';
 import Certificate from '../../crypto_wrappers/x509/Certificate';
 import { MockKeyStoreSet } from '../../keyStores/testMocks';
-import { issueGatewayCertificate } from '../../pki';
+import { issueGatewayCertificate } from '../../pki/issuance';
 import { StubPayload } from '../../ramf/_test_utils';
 import { SessionKey } from '../../SessionKey';
 import { NodeError } from '../errors';
