+++ src/lib/pki/issuance.spec.ts
@@ -1,16 +1,16 @@
 import * as pkijs from 'pkijs';
 
-import { generateStubCert } from './_test_utils';
-import { generateRSAKeyPair, getPublicKeyDigestHex } from './crypto_wrappers/keys';
-import BasicCertificateIssuanceOptions from './crypto_wrappers/x509/BasicCertificateIssuanceOptions';
-import Certificate from './crypto_wrappers/x509/Certificate';
+import { generateStubCert } from '../_test_utils';
+import { generateRSAKeyPair, getPublicKeyDigestHex } from '../crypto_wrappers/keys';
+import BasicCertificateIssuanceOptions from '../crypto_wrappers/x509/BasicCertificateIssuanceOptions';
+import Certificate from '../crypto_wrappers/x509/Certificate';
 import {
   DeliveryAuthorizationIssuanceOptions,
   GatewayCertificateIssuanceOptions,
   issueDeliveryAuthorization,
   issueEndpointCertificate,
   issueGatewayCertificate,
-} from './pki';
+} from './issuance';
 
 let stubSubjectKeyPair: CryptoKeyPair;
 let stubCertificate: Certificate;
